backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.3.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
requests==2.31.0
httpx==0.27.0
